Refactor and add Comments for Module Parameters
Seperate the module parameters into a sperate file to help wiht readability. Comments have also been added to each module parameter funciton to help wit understanding its functionality
